Move validate output to output generation
