+++ generate_output.py
@@ -4,8 +4,6 @@ import json
 import os
 import re
 
-from validate_output import validate_output
-
 TMP_DIR = "./tmp/"
 
 # Set up logging
@@ -42,6 +40,43 @@ class Subtitle:
         start_str = self.time_to_str(self.start)
         end_str = self.time_to_str(self.end)
         return f"{self.index}\n{start_str} --> {end_str}\n{self.text}\n"
+
+def validate_output(path):
+    """
+    Validate the provided path depending on whether it's an SRT file or a directory.
+
+    :param path: Path to the SRT file or directory.
+    :type path: str
+    :return: True if the destination location exists (and warns about potential overwrite); False if output directory doesn't exist.
+    :rtype: bool
+    """
+    if not path:
+        raise ValueError("Output path must not be empty")
+
+    # Check if path is a directory or a file
+    if os.path.isdir(path):
+        # The path is a directory; check if it exists
+        if os.path.exists(path):
+            return os.path.join(path, "output.srt")
+        else:
+            raise Exception(f"Output directory does not exist: {path}")
+    else:
+        # Expecting the path to be an SRT file from this point onwards
+        if not path.lower().endswith('.srt'):
+            raise ValueError("Invalid output file type. Please provide a path to an '.srt' file.")
+
+        # Determine the output directory
+        output_directory = os.path.dirname(path)
+
+        # Check if the output directory exists
+        if not os.path.exists(output_directory):
+            raise Exception(f"Output directory does not exist: {path}")
+
+        # Check if the file itself exists
+        if os.path.exists(path):
+            logging.warning(f"The file {path} already exists and will be overwritten.")
+
+        return path
 
 def convert_to_srt_time(time_in_seconds):
     formatted_time = datetime.timedelta(seconds=time_in_seconds)
