+++ modules.py
@@ -5,5 +5,4 @@ from chronometer import Chronometer
 from validate_audio_file import validate_audio_file
 from convert_hhmmss_to_ms import convert_hhmmss_to_ms
 from format_ms_duration import format_ms_duration
-from validate_output import validate_output
 from execution_args import execution_args
